% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_tippecanoe_version}
\alias{get_tippecanoe_version}
\title{Get Tippecanoe version}
\usage{
get_tippecanoe_version()
}
\value{
A character string with the Tippecanoe version or NULL if not installed
}
\description{
This function returns the version of Tippecanoe if it is installed.
}
\examples{
\dontrun{
get_tippecanoe_version()
}
}
