% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_tippecanoe_installed}
\alias{is_tippecanoe_installed}
\title{Check if Tippecanoe is installed}
\usage{
is_tippecanoe_installed()
}
\value{
TRUE if Tippecanoe is installed, FALSE otherwise
}
\description{
This function checks if Tippecanoe is installed and available in the system path.
}
\examples{
\dontrun{
is_tippecanoe_installed()
}
}
