% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm_version.R
\name{pm_version}
\alias{pm_version}
\title{Get PMTiles CLI version}
\usage{
pm_version()
}
\value{
Character string containing version information.
}
\description{
Display the version information of the PMTiles command-line tool.
}
\examples{
\dontrun{
pm_version()
}

}
