% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_pmtiles.R
\name{pmtiles}
\alias{pmtiles}
\title{Simple PMTiles creation}
\usage{
pmtiles(input_file, output_folder = "./pmtiles", output_name = NULL, ...)
}
\arguments{
\item{input_file}{Path to the input GeoJSON file}

\item{output_folder}{Path to the output folder (default: "./pmtiles")}

\item{output_name}{Name of the output file without extension (default: derived from input filename)}

\item{...}{Additional parameters to pass to create_pmtiles()}
}
\value{
The exit code of the Tippecanoe command
}
\description{
A simplified interface for creating PMTiles from GeoJSON files.
}
\examples{
\dontrun{
pmtiles("my_data.geojson")
pmtiles("my_data.geojson", attribution = "My Organization")
}
}
