% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmtiles-package.R
\docType{package}
\name{pmtiles-package}
\alias{pmtiles-package}
\title{pmtiles: Create PMTiles with Tippecanoe in R}
\description{
An R package that provides a simple interface to create PMTiles using the Tippecanoe command-line tool. PMTiles is a file format for efficiently storing map tiles as a single file.
}
\author{
\strong{Maintainer}: Zhao Wang \email{wangzhao0217@gmail.com} (\href{https://orcid.org/0000-0002-4054-0533}{ORCID})

Authors:
\itemize{
  \item Robin Lovelace \email{rob00x@gmail.com} (\href{https://orcid.org/0000-0001-5679-6536}{ORCID})
}

}
\keyword{internal}
